requirements.txt
# Core AI/ML Dependencies
langchain==0.1.0
langchain_community==0.0.10
langchain_core==0.1.10
langchain_groq==0.1.9
langchain_openai==0.0.8
langchain-google-genai==1.0.1
langgraph==0.0.69
faiss-cpu==1.7.4

# Web Framework Dependencies
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Security and Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Connectors and Integrations - GitHub, Jira, Slack
requests==2.31.0
httpx==0.25.2
boto3==1.34.0
github3.py==4.0.1
atlassian-python-api==3.41.10
slack-sdk==3.26.1
jira==3.5.2
PyGithub==2.1.1

# Data Processing and Analytics
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0

# Utility Dependencies
loguru==0.7.2
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1
aioredis==2.0.1
typing-extensions==4.8.0
click==8.1.7
rich==13.7.0
